Benign bug fix: yynt-initialize-cache always creates a new table.
When comparing cache-items of the PROJECT with the fields of the database table,
the database table has a path field, which cache-items cannot have. Therefore,
the field equality check always fails, resulting in the creation of a new table.

However, since the new table will copy fields with the same name from the old
table, the original information will not be lost despite the creation of a new
table.